significantly improve the performance of the EnemyCasts toggles
there are a few tricks here:

- applying `React.memo` to some child components to prevent re-renders
  when the parent re-renders due to a state change but doesn't impact
  their props
- pushing state as far down as possible (moving props off of the
  TimelineTab / TimelineComponent components and into EnemyCasts so that
  they don't trigger re-renders of other content)
- using a `ref` to toggle the css variables instead a `style` prop to
  limit the amount of work that has to be done in response to clicking a
  toggle. (this may no longer be necessary with the other two, but I saw
  no reason to remove it)
